move oso_mcp_tools
